Fix SyntaxGraph annotation being a typedef
Summary:
There have been a few bugs around structured annotation handling in `SyntaxGraph` and it reolves around the inconsistent way in which annotations are captured in `schema.thrift.`

All other types in `schema.thrift` use `TypeStruct` when referring to other types. However, annotations use either URI or `DefinitionKey`. In either case, we have to write some "special" code to deal with this incidental difference, which leads to bugs.

The change here coerces `DefinitionKey` into `TypeStruct` so that the actual type resolution code is shared with the rest of `SyntaxGraph` implementation. This means we'll be using much better tested code and hopefully there are no more bugs around this.

The "proper" fix might be updating `schema.thrift` to use `TypeStruct` and also actually enforcing in the compiler that unions, typedefs to unions etc. are not used as annotations.

Reviewed By: pranavtbhat

Differential Revision: D71662647

fbshipit-source-id: 5e7f26c8f05b16b337f328ea1df829cd87af6e61