implement basic field access
Summary:
This diff implements `from_python` for unions and structs as well as property getters. The basic idea is we can defer all data storage to an inner thrift-python object, making conversion very cheap. However, when accessing a container or structured attribute, we do have to do a cheap conversion on that. Hence, we cache those fields that require conversion.

We define `__slots__` to prevent users assigning random attributes, reduce memory footprint, and maintain same interface as cython, which disables class `__dict__`.

Reviewed By: prakashgayasen, Sunnie912

Differential Revision: D70823034

fbshipit-source-id: 529fed5fb1b21d4b9d12cbe11b39edfc6a711d77