add serialization support
Summary:
This diff implements serialization. It's a bit awkward because the best place to do it is in the `serializer` lib, so use `_fbthrift__inner` attribute, and `_FBTHRIFT__PYTHON_CLASS` to signal whether thrift-py3 has in-place migrate enabled.

There may be a marginal perf cost to this on trunk, but it's negligible compared to the cost of serialization itself. Moreover, we prioritize thrift-py3 to thrift-python migration over perf of thrift-py3 itself.

Reviewed By: prakashgayasen, Sunnie912

Differential Revision: D70823029

fbshipit-source-id: 3b25a39e09c6ece1a33160c8ae50122f907e4761