Use MethodHandle to acquire ForkJoinPool constructor
Summary:
Java 11 introduced a new constructor the the ForkJoinPool that allows us to configure the internal pool constraintes. We would like to control the maximumPoolSize. In some scenarios, such as blocking calls, it’s possible that the ForkJoinPool will create up to MAX_CAP (0x7fff, 32767) threads to ensure the pool is making progress. This creates a scenario where the host becomes unreponsive due to context switching overhead.

This diff ensure that in Java 11+ we configure the max pool size to avoid unbounded thread growth and to ensure that hosts don’t become unhealthy. To do this we need to detect the java version and then use MethodHandle look up to identify the constructor we expect. In all cases Java 11+ we’ll use the new advanced constructor and cap the pool size at the target parallelism.

Reviewed By: prakashgayasen

Differential Revision: D71574956

fbshipit-source-id: eef40ed2f5a840f5da4c93ce7e704a388fc72638