enforce no inheritance from generated thrift
Summary:
In thrift-py3-trunk, we rely on `cython.final` to ensure no inheritance. In thrift-python, we use the `_fbthrift_allow_inheritance_DO_NOT_USE` class attribute plus some logic in the `StructMeta` to ban inheritance unless this class attribute is present.

To make this work in thrift-py3-inplace, we have to adapt the same strategy as thrift-python. This still works on trunk because we use `_FBTHRIFT__PYTHON_CLASS` to skip enforcement.

Reviewed By: prakashgayasen, Sunnie912

Differential Revision: D71586264

fbshipit-source-id: 506ec39a6bc811bfe601f92f557356bcd439e735