add fast path for python to py3 conversion
Summary:
Add a faster path for thrift-python to thrift-py3 conversion.

This improves performance relative to the reflection based `to_py3_struct` method called by `_to_py3(self)` in generated thrift-python types.

It's also a workaround for lack of support for converting `GeneratedError` subclasses, which is not implemented, but conversions of exceptions are required for in-place migrate.

Reviewed By: Sunnie912

Differential Revision: D71634927

fbshipit-source-id: 042c0ce65c6b35a8f540a022beb701132d99095a