Allow @rust.Arc to override @thrift.Box
Summary:
The `thrift.Box` annotation enables boxed semantics across a set of languages where otherwise flat cyclic data structures are not allowed. This includes C++ and Rust.

In fact for C++ there is no other endorsed C++-specific boxing annotation. There is `cpp.Ref{type = cpp.RefType.Unique}` which is similar but deprecated in favor of `thrift.Box`.

https://www.internalfb.com/code/fbsource/[ee00ed18a9c44330a1b72a6528b4ce8d024d1ffb]/xplat/thrift/compiler/sema/standard_validator.cc?lines=924-927

So if someone wants a field to be `std::unique_ptr` in C++ while being `Arc` in Rust, the only way to accomplish that was with the deprecated `cpp.Ref` annotation.

This diff removes the unnecessary conflict between `thrift.Box` and `rust.Arc`. The semantics of this pair of annotations is obvious enough. In Rust, the field will be `Arc`, while in non-Rust languages that respect `thrift.Box`, it will be boxed.

A field that has both `rust.Box` and `rust.Arc` will continue to produce a compile error.

Reviewed By: Imxset21

Differential Revision: D71859443

fbshipit-source-id: d9fe45949b8d20aa3a65e9328abf163f0c2ff582