Make regression test actually work
Summary:
We have a regression test for field name conflict with common template parameter names such as `T`. Unfortunately it doesn't work and as a result we recently missed a breakage when the pika toolchain was upgraded: https://www.internalfb.com/intern/test/844425131560812 was succeeding after D70140830.

Fix it by actually consuming the generated code.

Reviewed By: praihan

Differential Revision: D71989905

fbshipit-source-id: abbf0a237645169bb97658deca58bd64cf5255a1