Add support for '_' characters in thrift method names in generated service handler
Summary:
The Java Generated Service Handler has a binding where the method name is read from the RPC metatdata, that name is used to map the request to the given implementation. The implemenations are always use manged Java based names that are cammel case, to conform to language norms. The mapping however looked upt he "java" name and not the thrift name which could result in the look up for the particular method to fail. This diff fixes this to allow the thrift name lookup to be performed.

Example method delegation:

```
 public Mono<ServerResponsePayload> singleRequestSingleResponse(ServerRequestPayload _payload) {
        String _name = _payload.getRequestRpcMetadata().getName();

        ...

            Mono _result;
            try {
                switch (_name) {
                    case "direct_publish_client_to_mqtt":
                        _result = _dodirectPublishClientToMqtt(this._delegate, _payload, this._directPublishClientToMqttReaders, _chain);
                        break;
        ...
```

Diff showing before and after, (includes MQTT thrift change) https://www.internalfb.com/intern/diffing/?paste_number=1767499993

Reviewed By: prakashgayasen

Differential Revision: D71933352

fbshipit-source-id: 0b04dd6ced7d6104b5739770ca3add23829ed7e0