add fixture showing in-place + auto-migrate
Summary:
Auto-migrate generally supersedes in-place migrate, in the sense that auto-migrate causes no thrift-py3 types to be generated. Since in-place migrate changes how thrift-py3 types are generated, that means in-place trivially has no effect when auto-migrate is enabled.

The one exception to this is in the thrift-py3 converter module, where we have to be more careful since the `.converter` functions are generated even when auto-migrate is turned on. Still, auto-migrate supersedes in-place in this case.

Reviewed By: prakashgayasen

Differential Revision: D72176772

fbshipit-source-id: c4c6c63cacb492b97ab6598e11994eef0936510b