python 3.12 coverage regression workaround
Summary:
https://github.com/python/cpython/issues/107674 slowdown in 3.12 settrace behavior this affects coverage

https://github.com/python/cpython/issues/127953 quadratic behavior for dictionaries.
also possible
https://github.com/python/cpython/pull/107780 which has to do with boxing and unbox of python ints

These fixes have not been merged into 3.12 trunk, for one the quadratic behavior fix doesn't patch cleanly.

Enums are particularly painful because in thrift-py the .py contains two potentially large dict literals and a large class declaration.   The problem was less pronounced in thrift-python because it doesn't have these large dicts and there seems to be some optimization for enums

Removing just the dictionaries and leaving the class was not enough.
Removing the large class was not enough.

But  building up the dictionaries and class at import time solves the problem.

Reviewed By: itamaro

Differential Revision: D72197140

fbshipit-source-id: 2c209bb12f232a0001570b77ec0e9fd28fcf4965