Enforce global client interceptor name uniqueness
Summary:
This diff adds unique name enforcement for global client interceptors at Meta.

My justification is as follows: Though logging / killswitching for client interceptors in general might not make sense because the set is unknown to the library, this is not true for global client interceptors which have a higher blast radius and are generally not optional. In order to do this, global client interceptors would need some unique identifier - I propose using the interceptor name.

Reviewed By: praihan

Differential Revision: D71989537

fbshipit-source-id: 0294860a1f9ce3ccc2f655867f5335b6b5eb88da