Add NativeObject unit tests
Summary:
This diffs adds the skeleton for a set of unit tests that verify the correct behavior of `experimental::Object`.

## Notes

This stack of diffs has been split up to aid reviewing for each sub-area, e.g. object, containers, ...
The addition of some functionality might only make sense when looking at the entirety of the stack, e.g. `ValueHolder`

Reviewed By: thedavekwon

Differential Revision: D72347750

fbshipit-source-id: 118ec78a5af786ed0a900b4e0c76004ac8d4ef5c