accommodate iterable value for list field
Summary:
Some thrift-py3 usage gives a `map` as argument to a `list` field. This succeeds in thrift-py3 and fails in thrift-python.

To avoid unexpected breakage during migration, we should accomodate this for now because we can do so with minimal performance hit, since `try` is basically free in non-exception case if you ignore the effect on instruction cache.

Reviewed By: prakashgayasen

Differential Revision: D72606617

fbshipit-source-id: ea13e8d6840e3562a89b2e494de5275c5602a387