Fix ResetStateTest on macOS
Summary:
`ResetStateTest` relies on Linux-specific behavior of `connect` which fails quickly if the server socket is not listening. We still get an asynchronous event since we connect is non-blocking mode but it's really fast so doing 1000 iterations works.

On macOS `connect` timeouts in this case which doesn't work with large number of iterations and likely exercises the wrong path.

Fix this by passing and invalid socket address instead of a non-listening one. This still communicates the failure through an asynchronous event (as opposed to eagerly failing `connect`) so it exercises the same failure path and is consistent between platforms.

Reviewed By: praihan

Differential Revision: D72729747

fbshipit-source-id: 3c1da1f329055b653dbefb86e188d3f5d6b3c915