Fix race condition
Summary:
Users may be using the same "context" object for multiple parallel Thrift requests.

With the old implementation of `unionMaps` - multiple Go routines could try to write into the request-headers map concurrently - causing a panic:

https://www.internalfb.com/code/fbsource/[95ff23dcbed8]/fbcode/thrift/lib/go/thrift/rocket_client.go?lines=160-161

https://www.internalfb.com/code/fbsource/[8bfeac2dc66e]/fbcode/thrift/lib/go/thrift/rocket_client.go?lines=218-227

To fix this problem - we should not write into request headers maps at all - we should be creating a brand new map and copying the contents of request headers and persistent request headers into it.
This way we avoid concurrent writes and are only left with concurrent reads - which are totally safe.

Reviewed By: mahmednabil109

Differential Revision: D72889233

fbshipit-source-id: 6124590b050bd258e9016d7165ce17d29611691e