Add NativeList support
Summary:
This diff introduces `NativeList`, a tagged representation of various std::vector based specializations of `Value`.
In the `Protocol::Value` implementation a list was always represented as `std::vector<Value>`, significantly increasing the memory use & reducing cache efficiency for primitive types.

With `NativeList` we provide specializations for lists of all primitives, e.g. `list<i32>`, `list<string>`, ...
so that consumers opt-in to the native std::vector equivalent, at the cost of resolving a single external tag for each list (in contrast to a tag for each element).

## Note

This is a common use-case in performance sensitive usage of dynamic thrift operations, e.g. operations on tensors/large lists of numbers

Reviewed By: thedavekwon

Differential Revision: D72348742

fbshipit-source-id: 291ba88b36fe9a9027643e213a9e32adb71460cf