Create bundle for thrift/lib/thrift/*
Summary:
Just like with annotations, this diff creates a C++ library which bundles the source code for `thrift/lib/thrift/*`.

This satisfies the requirement in `t_ast_generator` for `schema.thrift` (and transitives) to always exist. With `thrift2ast` becoming more prevalent with Dynamic Thrift, we want to make sure that its schematizer works without depending on the filesystem. For example, we expect that `Compiler.ligen` will soon compile and output `schema.thrift` directly rather than dumping to the filesystem.

Reviewed By: iahs

Differential Revision: D72480706

fbshipit-source-id: a7d524544f1145d671ce0333da909ac335cac66f