Use the same Thrift opt out list as PZ
Summary:
The XSU consent service is used to determine whether users data can cross universes and is necessary for supporting our privacy commitments.
Calling LoggerProductFiltering::genEffectivePolicyProduct loads user data and for XSU to know whether the data load should be allowed it looks up consent, which calls the consent service and triggers an infinite recursion.

I want to stop calling the IPF evaluation for core operational services (just like we do for PZ)

In this diff we share the code with PZ.

Reviewed By: reedriley

Differential Revision: D72995999

fbshipit-source-id: e50f916e1b3a4c184334f8b7c2c50a21fef362db