2/4[thrift][resource pools] Unsync `maxQps` from resource pools when `executionRate` is explicitly set.
Summary:
This diff enables services to turn off syncing of ThriftServer's `maxQps` to ConcurrencyController's `qpsLimit` (`maxQps` syncing), and turn on syncing of ThriftServer's `executionRate` to ConcurrencyController's `qpsLimit` (`executionRate` syncing). This is achieved by explicitly calling `ThriftServer::setExecutionRate`.

This is part of a migration to decouple ThriftServer's `maxQps` from ConcurrencyController's `qpsLimit`. The rationale is that `qpsLimit` should only apply to the loadshedding behavior in ThriftServer, and not also to the qps-limiting behavior of resource pools. The newly-added `executionRate` will instead apply to ConcurrencyController's qps-limiting behavior.

`maxQps` syncing remains on by default until `ThriftServer::setExecutionRate` is called. In later diffs, once all services that are dependent on the old `maxQps` syncing behavior have been migrated to use `executionRate` syncing, `maxQps` syncing will be removed and `executionRate` syncing will be always-on.

Meta:

This diff was derived from D68283214. The following variables/mechanisms were adapted for this diff.
- `maxRequests` -> `maxQps`
- `executionLimitRequests` -> `qpsLimit`
- `concurrencyLimit` -> `executionRate`

Reviewed By: sazonovkirill

Differential Revision: D72768208

fbshipit-source-id: 2f1cdfd66145e3d5bcbd0a2269edcaaccc1178e3