tests document non-compliant field access return
Summary:
Tests in thrift-python that document incorrectly typed field values. For `int` fields, it accepts `enum` value on init or call, then returns the un-modified enum when field is accessed. Given that generated enums have `int` base, an implicit conversion to `int` is probably fine, but the returned enum value from the `int`-typed field is arguable incorrect.

While this may be considered pyre-compliant, it's not thrift spec compliant:
- the return value the field changes before and after it's round-tripped through serialization
- it's giving py-deprecated vibes

Reviewed By: prakashgayasen

Differential Revision: D73066351

fbshipit-source-id: e6fafe97da048d1ba38870231bd64584e48fef72