use the thrift-py3 __Type enum in unions
Summary:
The `Union.type` and `Union.Type` api was left partially unimplemented in the hopes we could easily unify thrift-py3 and thrift-python enums on trunk.

Since that appears to be a non-trivial lift, add a test documenting current behavior and update inplace-migrate to have self-consistent behavior that matches py3 trunk.

Reviewed By: createdbysk

Differential Revision: D73063538

fbshipit-source-id: a8774af8909938d685712075505d366bcebd54dc