int field always returns `int`
Summary:
The thrift-python `IntTypeInfo` currently accepts any instance of `int` on python `__init__` or `__call__` for setting a thrift `i16|i32|i64` field. On accessing the set field, the thrift-python struct returns the *original* Python object, which is not necessarily an `int`. This violates the thrift spec, especially because the object's state changes across a roundtrip serialization, as documented in tests.

This diff continues to accept instances of `int` subclass when creating the struct from Python, but it will immediately convert it to `int` for thrift-spec compliance.

Reviewed By: aristidisp, createdbysk

Differential Revision: D73014965

fbshipit-source-id: d185c0ba59c556c52690d718fbed0e45e1cfa3fc