Optimized version of making Intercompatibility for Py3 Structs by default and also clean up dead enum related logic.
Summary:
The original version D73222710 was reverted because it significantly increased the build time. Upon investigation, it was hypothesized and confirmed that the issue was largely caused by compile time that scaled super-linearly with the number of branches in the `__init__` method of the `Union` class in Python 3.

This diff relocates the logic to a helper function to decrease the build time.

Reviewed By: ahilger

Differential Revision: D73549855

fbshipit-source-id: 8b264f0c8afbf88675434306f2c297c1904038d6