exception intercompatibility
Summary:
Thrift-python already supports inter-compatibility of thrift-py3 structs and unions, where they can be used to initialize a thrift-python field. For consistency, we should also support thrift-py3 GeneratedError subclasses.

This is also an opportunity to clean-up the logic and make it more Pythonic, with a `try`-`except AttributeError` to replace the base class `isinstance` check.

Reviewed By: Sunnie912

Differential Revision: D73552263

fbshipit-source-id: a97b415fe46408cee4190a83f2c4b39dc74c9b23